Modified application to use SDV
